requirements.txt
# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# LangChain & LangGraph
langchain>=0.1.0
langchain-community>=0.0.10
langchain-text-splitters>=0.0.1
langgraph>=0.0.20

# PDF Processing
pypdf>=3.17.0

# Embeddings & Vector Store
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
numpy>=1.24.0

# LLM Providers
google-generativeai>=0.7.2
httpx>=0.25.0

# Redis
redis>=5.0.0

# Observability
langsmith>=0.0.70

# Utilities
requests>=2.31.0
